requirements.txt
streamlit
langchain
langchain-groq
langchain-community
python-dotenv
pypdf
beautifulsoup4
youtube-transcript-api